Refactor the UI class
