+++ src/main/java/bhavs/Bhavs.java
@@ -185,6 +185,7 @@ public class Bhavs {
      * Runs the chatbot, initializing the UI and loading tasks from storage.
      */
     public void run() {
-        ui.loadTasks();
+        // ui.getCommandProcessor().
+        ui.getTaskManager().loadTasks();
     }
 }
